UPDATE I have finished this landing project. I added the last section of the page and finished the CSS requirements Odin gave me.
For this commit I will keep the borders visible so you can see my work and how I managed to align everything, the next commit everything will remain the same but the borders will no longer be visible. To be honest I struggled a bit with this project, it forced me to really understand the box method and how to use flexbox to your advantage. As I was building the page it was getting more and more difficult to keep track of all the <divs> so I decided to build the page in portions: Header, Top, Middle, Bottom, and Call to action. Odin recommends completely building out the HTML first, but that just lead me to confusion and requiring me to scrap everything and do it again. Almost all of the elements have flex so when you change the window size everything should move together, the issue arises when you make the window too small. The format begins to break down and things can look a little wonky. I believe min-width min-height can help prevent this, but I will need to look more into it. I also had trouble making he .call-to-action background color rounded. I tried to use the border-radius but it was not acting as it did with the buttons and images, this is something I will also look at. I feel as I am done for now, but when I learn a bit more, I plan on coming back to this project to make it cleaner and more appealing.

thanks for reading.